requirements.txt
aiodns
aiohappyeyeballs
aiohttp
aiosignal
altgraph
attrs
backtrader
blinker
ccxt
certifi
cffi
charset-normalizer
click
cloudpickle
coincurve
colorama
contourpy
cryptography
cycler
dash
dash-bootstrap-components
dask
filelock
Flask
fonttools
frozenlist
fsspec
idna
importlib_metadata
iniconfig
itsdangerous
Jinja2
joblib
kiwisolver
locket
markdown-it-py
MarkupSafe
matplotlib
mdurl
mpmath
multidict
narwhals
nest-asyncio
networkx
numpy
packaging
pandas
partd
pefile
pillow
plotext
plotly
pluggy
polars
polars-runtime-32
propcache
pyarrow
pycares
pycparser
Pygments
pyinstaller
pyinstaller-hooks-contrib
pyparsing
PyQt6
PyQt6-Charts
PyQt6-Charts-Qt6
PyQt6-Qt6
PyQt6_sip
pyqtgraph
pytest
python-dateutil
python-dotenv
pytz
pywin32-ctypes
PyYAML
requests
retrying
rich
scikit-learn
scipy
setuptools
six
sympy
ta
threadpoolctl
toolz
torch
torchaudio
torchvision
typing_extensions
tzdata
urllib3
websocket-client
Werkzeug
yarl
zipp